Basic echo RAT
It implements very basic communication commands, and commands' parsing, of course.
For now, it has the problem of not terminating gracefully when run with kscldr.sys: it exits fine, but when kscldr does de-allocation stuff, it causes bug check.
It's not a problem of the generated shellcode itself, it's an integration problem between the shellcode and kscldr; I will evaluate if the problem has to be solved, by modifying something in the shellcode or something in kscldr driver, or not solved (which can be good because the shellcode will be loaded in a different way during adversary emulation).