Graceful termination of basic echo RAT
The termination is intended w.r.t. kscldr driver; to do it, I improved the code quality by some refactoring, in particular about the application buffers' handling.
Then I inspected many failures, and concluded that ExFreePoolWithTag must not be called on the two buffers using during the application, for some reason.
The return still have to be done in DISPATCH_LEVEL.
The loader driver calls ```IoCompleteRequest(Irp, IO_NO_INCREMENT);``` in ```scldrCreateClose```, that is synchronous without sleep. In DISPATCH_LEVEL, in some cases it may cause indefinite hang.
So, the shellcode terminates gracefully, but, sometimes, (maybe for some race conditions idk) the loader driver hangs forever.
It's okay as it is, clean enough.