Implemented some adversary actions for the rookit, added version's informations
ATT&CK tactics involved with the actions: Credential Access, Discovery, Collection, Exfiltration.

Note that Defense Evasion and Command&Control are not included because they're part of the rootkit itself, not something implemented with the commands.

The list of commands available at this commit are:

- echo <string>
- close
- version
- pslist
- dump <process_name>

There is also a fallback to "invalid command" if other commands are issued, with the exception of this command:

```
dir <directory_name>
```

That says "not implemented"; I tried to implement directory listing, by adding the crate ntfs and some code (now commented) in the main, but it has some problems (ZwQueryDirectoryFile doesn't fill the FileInformation buffer), so I will leave it commented for now.