Added first-stage shellcode
In this commit there is the distinction between first-stage kernel shellcode and second-stage; the first-stage has yet to be optimized, and the "egg" (which must be replaced by the hypervisor with the second-level shellcode) is for now the sequence "int 3; ret" for testing with the kernel debugger. It works, but to be useful it has to be made shorter, otherwise it can't replace something like the code of a system call.