Added compiled kernel shellcodes and utils
The shellcodes in /kernel_shellcode_library are the last ones compiled, with the working integration among first-stage and second-stage.
The first stage is not expected to change, whereas the second stage can be changed to add more capabilities.
Some utils were command line utils used to speed-up some operations, like changing the base address of ntoskrnl after reading it with volatility and inspecting shellcodes' assembly to look for the right offset to patch.
There are also the working_first_stage/ and working_second_stage/ sub-folders, with other compiled kernel shellcodes which "work", but they don't integrate well.
In fact, they were tested using Flare KSCLDR: https://github.com/mandiant/flare-kscldr
The archive with kscldr compiled-in is kernel_loader.zip, it is compiled for Win10 build 19041.