Skeleton for adversary emulation application
The application's service is called "k-caldera" because it wants to be similar to MITRE's Caldera (altough simpler), with the difference that it uses arbitrary read&write to VM's memory from hypervisor for Initial Access.
It assumes that each target domain's RAM is mapped under /dev/shm .
In fact it is dockerized and /dev/shm is a mount point.