Improved interactive command-line interface, fixed version management
The interactive command-line interface was improved, but not with argparse because it's interactive; it was improved using the dictionary pattern.
There was also added commands to modify some configurations parameter at run-time, and to show info about available commands.
Windows 10 documented version was wrong: it was 19044 (21H2), and not 19041 (2004); in addition to that, it was noted that minor version also affects the execution of the exploit:

- the injection point may have something different, this was handled by adding the possibility of having more than one injection point for each major.minor version and to have different expected bytes for each injection point, to add reliability (but one working injection point for each major.minor pair should be enough);
- functions' position independent addresses (i.e., their offsets w.r.t. ntoskrnl's base address) change with minor version, so they must be found for each major.minor pair, for example by looking at symbols while doing kernel debug with WinDbg.

So it was added the minor version as another configuration parameter.