C2 connections' persistence, web interface skeleton, text UI over network
The persistence of connections to C2 server was implemented by putting log files in a volume and by storing the connections object as a JSON file, using a wrapper class around it which accesses it in mutual exclusion.
The web interface will be informative, without a rich command interface, because for that it was easier and more practical to expose the text UI over network.
It was exposed with a forking TCP server, so it was used posix IPC for mutexes and shared queues, instead of thread objects; it is also more performant and global variables were removed.