Implemented read and write of regular files in second-stage shellcode
The related rootkit's commands were also implemented.
Files are opened synchronously, this makes also easier to implement the read of large files using a small buffer: multiple reads are done using the I/O manager to keep track of the offset, and there is a socket.send() call for each bucket of data.
For now, the write isn't done using unconstrained data from outside, but is done using the entire command line, just to test that it works.
Another thing done was refactoring of the method to build an unicode string object from a character array: it was refactored into ntdef::macros; the previous method is left commented (in this commit) to highlight the difference.