Implemented basic operations with Windows registry in second-stage shellcode
Namely, it was implemented the "open key" for both the cases of generic read and generic write, and the "query value key".
To implement the related command, it was compulsory to implement one of the two following things: parsing of more than one argument in the rootkit, or a protocol to complete a command with multiple receives.
The second approach was chosen, because it will allow the proper implementation of the "write" command, which could be used for example to overwrite auto-start DLLs.
So there was a little refactor of the main of the second-stage shellcode, with the implementation of a "recv_msg" function.