Proper implementation of write command in second-stage shellcode
Now it's possible to write arbitrary data to any file, with a simple protocol.
The first step of the protocol is to send a command "write <file_full_path>".
After that, if the rootkit is ready to receive data to write, it will reply "send data to write".
At this point the C2 server can send up to 1460 bytes that will be written:

- if it sends less then 1460 bytes, then the rootkit will conclude that this is the last packet, so it will write them and close the file, without sending "send data to write";
- if it sends exactly 1460 bytes, then the rootkit will conclude that there are other bytes to receive, so it will send "send data to write" and wait for other data (the offset in the file is handled by Windows I/O manager since the file is opened with the option FILE_SYNCHRONOUS_IO_NONALERT).

If the amount of bytes to be written is multiple of 1460 bytes, then the C2 server has to send one garbage byte to be written at the end of the file to terminate the protocol.