Implemented automation of commands
The automation of commands was done borrowing some ideas from MITRE's Caldera:

- an attacker's "Ability" was implemented as an "Action", that is composed by one or more commands;
- the attacker's behaviour is a list of actions, called "Operation";
- each action may require output from previous actions or provide input to following actions: this kind of data is called "Facts".

There were implemented an interface for facts (FactsStorage), after having improved the handling of the ConnectionStorage; an "abstract" interface for a generic action and for a generic operation, with documentation about the "Parser" approach to generate facts from commands' outputs; many different parts of the C2 server were improved, for example to reliably know the index of the command output after using c2_send; there was some refactoring of the data structure of the connection, to make its schema easier to find and document; various new functions have been implemented in the command-line interface; the Dockerfile has been re-ordered for build speed.

Actions and operations were both implemented synchronous.
Some concrete parsers, actions and operations were implemented, to provide examples.
A few parsers need to be fixed, and a function to list available facts for a connection needs to be implemented in the command-line interface.