Implemented web interface, actions with parameters and integration among the interfaces
The web interface has the following functionalities:

- allows to see the connection table;
- for each connection, allows to see the related files;
- allows to see files in an upload_dir;
- allows to upload files in the upload_dir.

The only action with parameters is 'write_file', which allows to specify (optionally) a file from the upload_dir to read and send to the agent (preparing the command list accordingly).
So we expect to have some pre-defined payloads in the upload_dir, for example some malicious DLLs, and there is also the possibility to add new ones by uploading them.
It was also implemented an 'ls' command in the command-line interface, to see the list of files in upload_dir.
For the 'read_file' action it was implemented a parser that exfiltrates the file to the related connection folder, so that it can be downloaded from the web interface.
For the generic C2Operation, it was added to the report the list of command indexes for each action of the operation.