Beta implementation of ring0-to-ring3 indirect technique
I called the technique "schtask" because it makes use of scheduled tasks.
It tries to start a scheduled task (with a registration trigger, so it is one-shot) which executes powershell.exe bypassing the execution policy and executing a script in a hard-coded location.
The agent is able to overwrite the script time-by-time and make the scheduled task start again.
It gets output data from files generated by the script (so the operation has to be well designed).

To reach this objective, it was compulsory to implement functions to write to the Windows registry, in particular:

- \Registry\Machine\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Schedule\TaskCache\Tree\<task_name>
- \Registry\Machine\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Schedule\TaskCache\Tasks\<task_uuid>

See [https://cyber.wtf/2022/06/01/windows-registry-analysis-todays-episode-tasks/](here) for more info.
For now, all the writes are hard-coded and made after receiving a "schtask" command. In future I plan to refactor them, by implementing a protocol for arbitrary writes to the registry.

In addition to that, I came into a bug where Volatility wasn't able to find PDB files, so it crashed during the injection. I added exception handling for that, and also added to the build files some pre-fetched PDB files for Windows; this fixed the error.