Refactored second-stage shellcode schtask command into a generic setkey command, resolved issue #2
Since schtask looked hard to implement working in a generic way that didn't need patching, and since it had a very high cyclomatic complexity, it was chosen to replace it with a "setkey" command which implements a protocol to receive key name, subkey name, subkey type and subkey value, calling ZwCreateKey and ZwSetValueKey accordingly.
Therefore, it was implemented a "set_registry_key" concrete action in Python. This one will be used in complex operations, and can be used to implement an high-level "schtask" concrete action which concatenates commands of different "set_registry_key" actions (it will be an action and not an operation because (1) it can be useful as a single piece in an operation and (2) it may need some run-time generation of values).

Issue #2 was resolved as stated in the related comment: by upgrading the version to release 2.0.1 .