Experimented schtask technique as C2 action
The implementation of the "schtask" technique, explained in a previous commit message, was done as a C2 action.
It allowed for more and faster tries then the implementation in Rust, because it wasn't necessary to change the shellcode.
The experiments highlighted the fact that, although the scheduled tasks' GUI shows the changes made in the registry, the actual task scheduler is not "notified" about it.
The hypothesis is that it loads all the tasks once at its startup, and then exposes a COM port to be notified about changes; therefore, it doesn't make "polling" on the system registry, neither on XML files.

So, the result of the research is that the "schtask" technique can only be used for persistence, not for execution from ring-0 to ring-3.
The action is still left partially implemented, also with the writing of the XML file for completeness, to ease further research on the topic.