Implemented portability for version 19044.2006, experimented another ring0-to-ring3 technique
The technique experimented is a raw swapgs-sysret sequence.
For that, it was implemented an "usermode" command in the second-stage shellcode and the related concrete action in Python.
The command receives an user-level shellcode, allocates for it an area, starts a thread using PsCreateSystemThread; the thread has as target a function to perform transition to user-mode and takes as parameter the memory in which the user-level shellcode is allocated, and will jump to it using sysret.
Some experiments were tried with it, but it goes bug check everytime on the sysret instruction.
This approach has the pro of being portable w.r.t. different Windows versions; other approaches, like the ones seen in rootkits, are hard to be made portable and sometimes hooked by EDRs, whereas the "indirect" techniques (such as schtask or "executable file replacement") are not very reliable.
So, in the future it would be good to make more research about this technique, because it would also be a nice extension to the capabilities of the kernel agent.
For now it's left as a comment, and the "usermode" command is implemented as stub.