Implemented the concept of iterative action
An iterative action is an action that has one or more facts expressed as a list of values.
Each command in which the fact is a list is replicated, performing the substitution of the fact name with each value of the list (that's the "iteration" part).
This is useful even for simple operations like listing a directory and reading all files in that directory.