Implemented "dump_lsass_process" concrete action and the related parser
The parser uses code similar to the one written by zerosum0x0 for SassyKitdi to generate a Windows minidump from the received process dump.
The idea then was to use pypykatz to extract credentials, but there are errors while parsing the generated dump, so the export is not perfect.
Anyway, for the adversary emulation the important thing is to perform remote actions to generate logs on the victim machine, not to have an excellent parsing logic on C2 side.