Implemented "persistence and credential access" operation, renamed "desktop discovery" to "desktop exfiltration"
The "persistence" part of the operation uses install_schtask, which was modified to use a trigger at boot (which still doesn't work for now, but it's okay); it also copies the powershell script executed by the scheduled task to the right location.
The "credential access" part dumps lsass.exe memory to a minidump.