Implemented "ransomware" operation
It has the following tactics: discovery, collection, exfiltration, impact.
It also leaves a message on the victim's desktop.

For the implementation, it needed a "special" behaviour in actions, which in theory should have been implemented using a sub-class of C2Action specific of the action, but to be fast it was decided to implement it as an edge case in the performAction method.