Fixed issue #6, updated docs
Issue #6 was solved by closing the posix queue after receiving "close" command on the same queue, and the case of overloading is handled by immediately sending a "close" to the connecting agent and writing an error log in the hello_log.

New docs contain a package diagram of the kernel shellcode library and documentation for the new campaign in the experiments.